requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.28.0
